In Dockerfile
Put the copy package.json to /app then run npm install.
The problem here is, "npm install" created a "node_modules", this folder is pretty heavy.
So we copy the package.json first, if nothing different, the "npm install" will execute very fast (load from cache). and after that it will copy the rest of the code to /app.